% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutTree.r
\name{cutTree}
\alias{cutTree}
\title{Cut a dendrogram}
\usage{
cutTree(ddr, K = 10)
}
\arguments{
\item{ddr}{A dendrogram object.}

\item{K}{Maximum number of clusters.}
}
\value{
clustering membership.
}
\description{
Cut a dendrogram
}
\details{
Automatically cut a dendrogram. The tree cut is starting is using fix height. Each time lower the height until two new clusters occur. Define the new clusters as active clusters. Stop until the new active clusters are not generated from the old active clusters.
}
\examples{
N0 = 10
G <- 1000
GDEp <- 50
GDEn <- 50
K = 4

studies <- NULL
set.seed(15213)
for(k in 1:K){
	astudy <- matrix(rnorm(N0*2*G),nrow=G,ncol=N0*2)
	ControlLabel <- 1:N0
	caseLabel <- (N0 + 1):(2*N0)

	astudy[1:GDEp,caseLabel] <- astudy[1:GDEp,caseLabel] + 2
	astudy[1:GDEp + GDEn,caseLabel] <- astudy[1:GDEp + GDEn,caseLabel] - 2

	alabel = c(rep(0,length(ControlLabel)),rep(1,length(caseLabel)))

	studies[[k]] <- list(data=astudy, label=alabel)
}


result <- biomarkerCategorization(studies,function_limma,B=100,DEindex=NULL,seed = 15213)
dissimilarity <- result$dissimilarity
ddr <- hclustTree(dissimilarity)
clusterMembership <- cutTree(ddr)
table(clusterMembership)
}
\author{
Caleb
}

