% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tightClustPam.R
\name{tightClustPam}
\alias{tightClustPam}
\title{Tight Clustering}
\usage{
tightClustPam(adist, target, k.min, alpha = 0.1, beta = 0.6, top.can = 7,
  seq.num = 2, resamp.num = 10, samp.p = 0.7, nstart = 1,
  remain.p = 0.1, k.stop = 5, random.seed = NULL)
}
\arguments{
\item{adist}{Input dissimilarity matrix.}

\item{target}{The total number of clusters that the user aims to find.}

\item{k.min}{The starting point of k0. See 'Details' for more information.}

\item{alpha}{The threshold of comembership index. Default value is suggested to be used.}

\item{beta}{The threshold of clusters stably found in consecutive k0. Default value is suggested to be used.}

\item{top.can}{The number of top (size) candidate clusters for a specific k0. Default value is suggested to be used.}

\item{seq.num}{The number of subsequent k0 that finds the tight cluster. Default value is suggested to be used.}

\item{resamp.num}{Total number of resampling to obtain comembership matrix. Default value is suggested to be used.}

\item{samp.p}{Percentage of subsamples selected. Default value is suggested to be used.}

\item{nstart}{Number of different random inital for K-means. Default value is suggested to be used.}

\item{remain.p}{Stop searching when the percentage of remaining points <= remain.p. Default value is suggested to be used.}

\item{k.stop}{Stop decreasing k0 when k0<=k.stop. Default value is suggested to be used.}

\item{random.seed}{If random.seed is NULL no random seed will be set. If random.seed is a number, it will be used as the random seed. This parameters should be used to get the same result for different runs.}
}
\value{
Returned value is a "tight.clust" object (list). The first element is the original data matrix. The second element is a vector representing the cluster identity (-1: scattered gene set; 1: the first cluster; 2: the second cluster; ...). The third element is a vector of the size of each tight cluster.
}
\description{
This function could perform the tight clustering algorithm using dissimilarity matrix as input.
}
\details{
Tight clustering method is a resampling-evaluated clustering method that aims to directly identify tight clusters in a high-dimensional complex data set and allow a set of scattered objects without being clustered. The method was originally developed for gene cluster analysis in microarray data but can be applied in any complex data. The most important parameter is k.min. A large k.min results in smaller and tighter clusters. Normally k.min>=target+5 is suggested. All other parameters do not affect the quality of final clustering results too much and are suggested to remain unchanged. This packages is directly inherited from R tightClust package
}
\examples{
N0 = 10
G <- 1000
GDEp <- 50
GDEn <- 50
K = 4

studies <- NULL
set.seed(15213)
for(k in 1:K){
	astudy <- matrix(rnorm(N0*2*G),nrow=G,ncol=N0*2)
	ControlLabel <- 1:N0
	caseLabel <- (N0 + 1):(2*N0)

	astudy[1:GDEp,caseLabel] <- astudy[1:GDEp,caseLabel] + 2
	astudy[1:GDEp + GDEn,caseLabel] <- astudy[1:GDEp + GDEn,caseLabel] - 2

	alabel = c(rep(0,length(ControlLabel)),rep(1,length(caseLabel)))

	studies[[k]] <- list(data=astudy, label=alabel)
}


result <- biomarkerCategorization(studies,function_limma,B=100,DEindex=NULL,seed = 15213)
dissimilarity <- result$dissimilarity
tightClustResult <- tightClustPam(dissimilarity, target=2, k.min=5)
}
\author{
Chi Song <song.1188@osu.edu>
}

