% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeDissimilarity.r
\name{visualizeDissimilarity}
\alias{visualizeDissimilarity}
\title{Visualize dissimilarity heatmap}
\usage{
visualizeDissimilarity(dissimilarity, ddr, clusterMembership, ...)
}
\arguments{
\item{dissimilarity}{Input dissimilarity matrix}

\item{ddr}{Input dendrogram}

\item{clusterMembership}{clusterMembership result from previous cut tree}

\item{...}{Other parameters for heatmap.2}
}
\value{
heatmap.2 object
}
\description{
Visualize dissimilarity heatmap
}
\details{
Visualize dissimilarity heatmap
}
\examples{
N0 = 10
G <- 1000
GDEp <- 50
GDEn <- 50
K = 4

studies <- NULL
set.seed(15212)
for(k in 1:K){
	astudy <- matrix(rnorm(N0*2*G),nrow=G,ncol=N0*2)
	ControlLabel <- 1:N0
	caseLabel <- (N0 + 1):(2*N0)

	astudy[1:GDEp,caseLabel] <- astudy[1:GDEp,caseLabel] + 2
	astudy[1:GDEp + GDEn,caseLabel] <- astudy[1:GDEp,caseLabel] - 2

	alabel = c(rep(0,length(ControlLabel)),rep(1,length(caseLabel)))

	studies[[k]] <- list(data=astudy, label=alabel)
}


result <- biomarkerCategorization(studies,function_limma,B=100,DEindex=NULL,seed = 15213)
dissimilarity <- result$dissimilarity
ddr <- hclustTree(dissimilarity)
clusterMembership <- cutTree(ddr)
visualizeDissimilarity(dissimilarity, ddr, clusterMembership)


}
\author{
Caleb
}

