% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeVariability.r
\name{visualizeVariability}
\alias{visualizeVariability}
\title{visualizeVariability}
\usage{
visualizeVariability(biomarkerCategorization, clusterMembership,
  groups = NULL, ...)
}
\arguments{
\item{biomarkerCategorization}{biomarkerCategorization object from biomarkerCategorization function}

\item{clusterMembership}{clustering membership}

\item{groups}{re-order the module order.}

\item{...}{extra parameters}
}
\value{
heatmap.2 object
}
\description{
visualizeVariability
}
\details{
visualizeVariability
}
\examples{
N0 = 10
G <- 1000
GDEp <- 50
GDEn <- 50
K = 4

studies <- NULL
set.seed(15213)
for(k in 1:K){
	astudy <- matrix(rnorm(N0*2*G),nrow=G,ncol=N0*2)
	ControlLabel <- 1:N0
	caseLabel <- (N0 + 1):(2*N0)

	astudy[1:GDEp,caseLabel] <- astudy[1:GDEp,caseLabel] + 2
	astudy[1:GDEp + GDEn,caseLabel] <- astudy[1:GDEp + GDEn,caseLabel] - 2

	alabel = c(rep(0,length(ControlLabel)),rep(1,length(caseLabel)))

	studies[[k]] <- list(data=astudy, label=alabel)
}


result <- biomarkerCategorization(studies,function_limma,B=100,DEindex=NULL,seed = 15213)
dissimilarity <- result$dissimilarity
ddr <- hclustTree(dissimilarity)
clusterMembership <- cutTree(ddr)
visualizeVariability(result, clusterMembership, labRow='',labCol='',na.color=par("bg"),lwid=c(0.1,4), lhei=c(0.1,15),margins = c(1, 1))

}
\author{
Caleb
}

