% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWFisher.R
\name{AWFisher.pvalue}
\alias{AWFisher.pvalue}
\title{AW}
\usage{
AWFisher.pvalue(p.values)
}
\arguments{
\item{p.values}{1}

\item{method}{1}

\item{log}{1}
}
\value{
1
}
\description{
R package for fast computing for adaptively weighted fisher's method
}
\details{
fast computing for adaptively weighted fisher's method
}
\examples{
n=154
p.values = matrix(rbeta(n*100000, 1,1), ncol=n)
res = AWFisher.pvalue(p.values)
hist(res$pvalues, breaks=40)
table(res$sum.weight)
pvalues=res$pvalues[order(res$pvalues)]
plot(-log10((1:NROW(pvalues))/(1+NROW(pvalues))), -log10(pvalues),xlab="theoretical quantile", ylab="observed quantile")
lines(c(0,100), c(0,100), col=2)
}
\author{
Caleb
}

